Create Karaoke Song Selector
Using 'AND' and 'OR' operators. 
Used queries to decide what songs to sing. For the first step, selected all the song titles.
Next step, added another SELECT that uses OR to show the titles of the songs that have an 'epic' mood or a release date after 1990.
Then, added another SELECT that uses AND to show the titles of songs that are 'epic', and released after 1990, and less than 4 minutes long.
Note that the duration column is measured in seconds.